feat: add individual delete and refactor

+++ client/src/pages/Advanced/components/Delete.jsx
@@ -2,7 +2,6 @@ import { useContext, useState } from 'react'
 import BoxFiltering from '../../../components/BoxFiltering'
 import AppContext from '../../../context'
 import { Button, Heading, Stack, useDisclosure } from '@chakra-ui/react';
-import ConfirmDialog from '../../../components/ConfirmDialog';
 import { useTranslation } from 'react-i18next';
 import { callAPI } from '../../../service';
 
@@ -16,18 +15,18 @@ async function deleteBoxes(filters) {
 
 export default function Delete() {
 	const { boxes } = useContext(AppContext);
-	const [filtered, setFiltered] = useState(boxes);
 	const [filters, setFilters] = useState([]);
-	const { isOpen, onOpen, onClose } = useDisclosure();
 	const { t } = useTranslation();
 
 	const handleDelete = () => {
-		deleteBoxes(filters)
-			.then((res) => {
-				alert(`${res.deletedCount} boxes deleted`);
-				window.location.reload();
-			})
-			.catch(console.error);
+		if (window.confirm(t('deletePrompt'))) {
+			deleteBoxes(filters)
+				.then((res) => {
+					alert(`${res.deletedCount} boxes deleted`);
+					window.location.reload();
+				})
+				.catch(console.error);
+		}
 	}
 
 
@@ -38,7 +37,7 @@ export default function Delete() {
 			<Heading>{t('delete')}</Heading>
 			<BoxFiltering
 				boxes={boxes}
-				setFilteredBoxes={setFiltered}
+				setFilteredBoxes={() => { }}
 				setFiltersOutside={setFilters}
 				includeProgress={false}
 				includeSearch={false}
@@ -46,16 +45,10 @@ export default function Delete() {
 			<Button
 				colorScheme='red'
 				variant='solid'
-				onClick={onOpen}
+				onClick={handleDelete}
 			>
 				{t('delete')}
 			</Button>
-			<ConfirmDialog
-				message={t('deletePrompt')}
-				onConfirm={handleDelete}
-				isOpen={isOpen}
-				onClose={onClose}
-			/>
 		</Stack>
 	)
 }
